addition of a treeview component
User is able to browse courses and tasks downloaded in the download file path given in extension settings. Login is required and treeview is wiped on logout. Task files can be opened to the editor by clicking, and multiple tasks can be opened at once using right-click-> open all tasks in this directory